refactor: extract navigation interface from header and move to navigation.ts
