+++ interfaces/header.ts
@@ -1,16 +1,9 @@
+import { INavigation } from '@interfaces/navigation'
+
 export interface IHeader {
   navigation?: Array<INavigation>
   logo?: ILogo
   theme?: any
-}
-
-export interface INavigation {
-  title: string
-  link: string
-  sections: Array<{
-    title: string
-    link: string
-  }>
 }
 
 export interface ILogo {
