+++ interfaces/navigation.ts
@@ -1,0 +1,11 @@
+export interface INavigation extends IBasicNavigation {
+  sections: Array<{
+    title: string
+    link: string
+  }>
+}
+
+export interface IBasicNavigation {
+  title: string
+  link: string
+}
